+++ src/three/gltf/classes/MeshFeatures.js
@@ -1,6 +1,6 @@
 import { Vector2 } from 'three';
-import { TextureReadUtility } from './utilities/TextureReadUtility.js';
-import { getTexCoord, getTexelIndices, getTriangleIndices } from './utilities/TexCoordUtilities.js';
+import { TextureReadUtility } from '../utilities/TextureReadUtility.js';
+import { getTexCoord, getTexelIndices, getTriangleIndices } from '../utilities/TexCoordUtilities.js';
 
 const _uv = /* @__PURE__ */ new Vector2();
 const _pixel = /* @__PURE__ */ new Vector2();
