+++ src/three/gltf/classes/StructualMetadata.js
@@ -1,0 +1,17 @@
+// TODO: there are cases where we create a Matrix or Vector, for example, and immediately
+// discard it due to how "noData" is handled
+
+export class StructuralMetadata {
+
+	constructor( definition ) {
+
+		// TODO
+
+	}
+
+	getAccessor( key ) {
+
+	}
+
+}
+
